requirements.txt
transformers[torch]
datasets
scikit-learn
accelerate
evaluate
torch
torchvision
openai-whisper
rich
sounddevice  # --mic mode; also needs system lib: Linux: sudo apt install libportaudio2 | macOS: brew install portaudio
# optional LLM summary providers (install whichever you use):
# pip install anthropic        (--llm-provider claude)
# pip install openai           (--llm-provider openai)
# pip install google-genai     (--llm-provider gemini)